登录和注册/requirements.txt
psycopg2
Flask
gunicorn
flask-debugtoolbar
Flask-WTF
Flask-SQLAlchemy
Flask-Login
WTForms-Components
